feat: consolidate TypeScript interfaces for all mock implementations
- Create centralized types.ts file with comprehensive interface definitions

- Standardize mock interfaces across dungeon-game-persistence, health-hooks, and storage-functions

- Add base mock interfaces with common functionality (reset, clear, restore)

- Include utility types for mock scenarios, factories, and options

- Remove duplicate interface definitions from individual mock files

- Ensure type safety and consistency across all mock implementations

- Update imports to use centralized type definitions

Related to #4.4
