+++ tasks/tasks-prd-codebase-cleanup.md
@@ -18,6 +18,7 @@
 - `__mocks__/dungeon-game-persistence.ts` - ✅ Centralized mock for dungeon game persistence with TypeScript interfaces and helper functions
 - `__mocks__/health-hooks.ts` - ✅ Centralized mock for health-related hooks with TypeScript interfaces and scenario support
 - `__mocks__/storage-functions.ts` - ✅ Centralized mock for storage functions with TypeScript interfaces and scenario support
+- `__mocks__/types.ts` - ✅ Centralized TypeScript interfaces for all mock implementations with comprehensive type definitions
 - `package.json` - Dependencies to be removed and updated
 - `README.md` - Documentation to be updated with new guidelines
 
@@ -67,7 +68,7 @@
   - [x] 4.1 Create **mocks**/dungeon-game-persistence.ts with centralized mock implementation
   - [x] 4.2 Create **mocks**/health-hooks.ts with standardized health hook mocks
   - [x] 4.3 Create **mocks**/storage-functions.ts with centralized storage mocks
-  - [ ] 4.4 Add TypeScript interfaces for all mock implementations
+  - [x] 4.4 Add TypeScript interfaces for all mock implementations
   - [ ] 4.5 Include helper functions for common mock scenarios (setup, teardown, reset)
   - [ ] 4.6 Update all test files to import from centralized mocks
   - [ ] 4.7 Define standard mock structure and naming conventions
